requirements-dev.txt
asgiref==3.5.2
dj-static==0.0.6
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
Pillow==9.3.0
pytz==2022.6
sqlparse==0.4.3
static3==0.7.0
